backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
deprecation==2.1.0
exceptiongroup==1.3.1
Flask==3.1.2
flask-cors==6.0.1
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.0
packaging==25.0
postgrest==2.24.0
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
realtime==2.24.0
requests==2.32.5
storage3==2.24.0
StrEnum==0.4.15
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.4
yarl==1.22.0
